bin/capture bin/git-to-changes.pl bin/playground-server bin/test-gpt
bin/capture now accepts a first arg of -, which causes it to run
the script with perl -- selected because perl respects shebang lines
and will run another interpreter is the script has a shebang that
does not point to perl.

bin/git-to-changes.pl now deletes the Changes file, to avoid
repetition.

bin/playground-server now adds "lib" to the libpath.  It still needs
work.

bin/test-gpt now has a mock mode which just opens a text editor and
lets you type in what you are pretending the AI said ... for things
like testing scripts.